Alpha 1.0
Specific Details will be listed in wiki page (to be made)
Brief Description:
The Roadkill app is now fully functional although the "settings" menu
item is yet to be implemented. 
Users can now save records and upload them to wildlifecrossing.net.
(Pictures are not supported at this time)
Records can be viewed as points on Google Map or in list form.
can view locations

TO DO:
-A lot of error checking (right now users can leave fields input fields
blank which will likely crash the app)
-Bug fixes
-Support for different server system